refactor(#3479): extract orphan status-panel cleanup from tmux_watcher.rs
Verbatim move (zero logic change) of the watcher-direct orphan status-panel
cleanup cluster — cleanup_orphan_external_input_status_panel,
complete_watcher_status_panel_v2, refresh_watcher_session_panel_from_lifecycle —
out of src/services/discord/tmux_watcher.rs into the sibling submodule
src/services/discord/tmux_watcher/orphan_status_panel_cleanup.rs (240 LoC). The
3 fns are pub(super) there and re-imported into the parent so the watcher loop's
call sites and the sibling single_message_footer.rs completion call stay
byte-identical. tmux_watcher.rs prod LoC 7241 -> 7022; ratchet baseline lowered
to match.
